Rename redux slice
