+++ src/pages/DetailsPage.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import { RiArrowGoBackLine } from 'react-icons/ri';
 import Details from '../components/Details';
-import { fetchFilms } from '../redux/films/Films';
+import { fetchFilms } from '../redux/films/FilmsSlice';
 import style from '../styles/Details.module.css';
 
 const DetailsDisplay = () => {
