+++ src/pages/FilmsPage.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import FilmsList from '../components/FilmsList';
-import { fetchFilms } from '../redux/films/Films';
+import { fetchFilms } from '../redux/films/FilmsSlice';
 import SearchBar from '../components/Search';
 import style from '../styles/FilmsPage.module.css';
