+++ src/redux/films/FilmsSlice.jsx
@@ -25,15 +25,21 @@ export const fetchFilms = createAsyncThunk(
   },
 );
 
-const initialState = [];
-
 const filmsSlice = createSlice({
   name: 'films',
-  initialState,
-  reducers: {},
+  initialState: {
+    films: [],
+    search: '',
+  },
+  reducers: {
+    searchUpdate: (state, action) => (
+      { ...state, search: action.payload }
+    ),
+  },
   extraReducers: (builder) => {
     builder.addCase(fetchFilms.fulfilled, (state, action) => action.payload);
   },
 });
 
+export const { searchUpdate } = filmsSlice.actions;
 export default filmsSlice.reducer;
